+++ pyteal/ast/__init__.py
@@ -36,9 +36,11 @@ from pyteal.ast.asset import (
 from pyteal.ast.acct import AccountParam, AccountParamObject
 from pyteal.ast.box import (
     BoxCreate,
+    BoxResize,
     BoxDelete,
     BoxExtract,
     BoxReplace,
+    BoxSplice,
     BoxLen,
     BoxGet,
     BoxPut,
@@ -169,6 +171,15 @@ from pyteal.ast.maybe import MaybeValue
 from pyteal.ast.multi import MultiValue
 from pyteal.ast.opup import OpUp, OpUpMode, OpUpFeeSource
 from pyteal.ast.ecdsa import EcdsaCurve, EcdsaVerify, EcdsaDecompress, EcdsaRecover
+from pyteal.ast.ec import (
+    EllipticCurve,
+    EcAdd,
+    EcScalarMul,
+    EcPairingCheck,
+    EcMultiScalarMul,
+    EcSubgroupCheck,
+    EcMapTo,
+)
 from pyteal.ast.router import (
     BareCallActions,
     CallConfig,
@@ -212,8 +223,10 @@ __all__ = [
     "BitwiseXor",
     "Block",
     "BoxCreate",
+    "BoxResize",
     "BoxDelete",
     "BoxExtract",
+    "BoxSplice",
     "BoxGet",
     "BoxLen",
     "BoxPut",
@@ -252,6 +265,13 @@ __all__ = [
     "EcdsaVerify",
     "Ed25519Verify_Bare",
     "Ed25519Verify",
+    "EllipticCurve",
+    "EcAdd",
+    "EcScalarMul",
+    "EcPairingCheck",
+    "EcMultiScalarMul",
+    "EcSubgroupCheck",
+    "EcMapTo",
     "EnumInt",
     "Eq",
     "Err",