+++ pyteal/ast/box.py
@@ -49,6 +49,49 @@ class BoxCreate(Expr):
 BoxCreate.__module__ = "pyteal"
 
 
+class BoxResize(Expr):
+    """Resize an existing box.
+
+    If the new size is larger than the old size, zero bytes will be added to the end of the box.
+    If the new size is smaller than the old size, the box will be truncated from the end.
+    """
+
+    def __init__(self, name: Expr, size: Expr) -> None:
+        """
+        Args:
+            name: The key used to reference this box. Must evaluate to a bytes.
+            size: The new number of bytes to reserve for this box. Must evaluate to a uint64.
+        """
+
+        super().__init__()
+        require_type(name, TealType.bytes)
+        require_type(size, TealType.uint64)
+        self.name = name
+        self.size = size
+
+    def __teal__(self, options: "CompileOptions"):
+        verifyProgramVersion(
+            minVersion=Op.box_resize.min_version,
+            version=options.version,
+            msg=f"{Op.box_resize} unavailable",
+        )
+        return TealBlock.FromOp(
+            options, TealOp(self, Op.box_resize), self.name, self.size
+        )
+
+    def __str__(self):
+        return f"(box_resize {self.name} {self.size})"
+
+    def type_of(self):
+        return TealType.none
+
+    def has_return(self):
+        return False
+
+
+BoxResize.__module__ = "pyteal"
+
+
 class BoxDelete(Expr):
     """Deletes a box given its name."""
 
@@ -83,7 +126,10 @@ BoxDelete.__module__ = "pyteal"
 
 
 class BoxReplace(Expr):
-    """Replaces bytes in a box given its name, start index, and value."""
+    """Replaces bytes in a box given its name, start index, and value.
+
+    Also see BoxSplice.
+    """
 
     def __init__(self, name: Expr, start: Expr, value: Expr) -> None:
         """
@@ -163,6 +209,68 @@ class BoxExtract(Expr):
 
 
 BoxExtract.__module__ = "pyteal"
+
+
+class BoxSplice(Expr):
+    """Splice content into a box."""
+
+    def __init__(
+        self, name: Expr, start: Expr, length: Expr, new_content: Expr
+    ) -> None:
+        """
+        Replaces the range of bytes from `start` through `start + length` with `new_content`.
+
+        Bytes after `start + length` will be shifted to the right.
+
+        Recall that boxes are constant length, and this operation will not change the length of the
+        box. Instead content may be adjusted as so:
+
+            * If the length of the new content is less than `length`, the bytes following `start + length` will be shifted to the left, and the end of the box will be padded with zeros.
+
+            * If the length of the new content is greater than `length`, the bytes following `start + length` will be shifted to the right and bytes exceeding the length of the box will be truncated.
+
+        Args:
+            name: The name of the box to modify. Must evaluate to bytes.
+            start: The byte index into the box to start writing. Must evaluate to uint64.
+            length: The length of the bytes to be replaced. Must evaluate to uint64.
+            new_content: The new content to write into the box. Must evaluate to bytes.
+        """
+        super().__init__()
+        require_type(name, TealType.bytes)
+        require_type(start, TealType.uint64)
+        require_type(length, TealType.uint64)
+        require_type(new_content, TealType.bytes)
+        self.name = name
+        self.start = start
+        self.length = length
+        self.new_content = new_content
+
+    def __teal__(self, options: "CompileOptions"):
+        verifyProgramVersion(
+            minVersion=Op.box_splice.min_version,
+            version=options.version,
+            msg=f"{Op.box_splice} unavailable",
+        )
+        return TealBlock.FromOp(
+            options,
+            TealOp(self, Op.box_splice),
+            self.name,
+            self.start,
+            self.length,
+            self.new_content,
+        )
+
+    def __str__(self):
+        return f"(box_splice {self.name} {self.start} {self.length} {self.new_content})"
+
+    def type_of(self):
+        return TealType.none
+
+    def has_return(self):
+        return False
+
+
+BoxSplice.__module__ = "pyteal"
 
 
 def BoxLen(name: Expr) -> MaybeValue: