+++ pyteal/ast/box_test.py
@@ -5,31 +5,37 @@ import pyteal as pt
 
 avm7Options = pt.CompileOptions(version=7)
 avm8Options = pt.CompileOptions(version=8)
+avm10Options = pt.CompileOptions(version=10)
 
-POSITIVE_TEST_CASES: list[Tuple[pt.Expr, pt.TealType]] = [
-    (pt.BoxCreate(pt.Bytes("box"), pt.Int(10)), pt.TealType.uint64),
-    (pt.BoxDelete(pt.Bytes("box")), pt.TealType.uint64),
-    (pt.BoxExtract(pt.Bytes("box"), pt.Int(2), pt.Int(4)), pt.TealType.bytes),
+POSITIVE_TEST_CASES: list[Tuple[int, pt.Expr, pt.TealType]] = [
+    (8, pt.BoxCreate(pt.Bytes("box"), pt.Int(10)), pt.TealType.uint64),
+    (8, pt.BoxDelete(pt.Bytes("box")), pt.TealType.uint64),
+    (8, pt.BoxExtract(pt.Bytes("box"), pt.Int(2), pt.Int(4)), pt.TealType.bytes),
     (
+        8,
         pt.BoxReplace(pt.Bytes("box"), pt.Int(3), pt.Bytes("replace")),
         pt.TealType.none,
     ),
-    (pt.BoxLen(pt.Bytes("box")), pt.TealType.none),
-    (pt.BoxGet(pt.Bytes("box")), pt.TealType.none),
-    (pt.BoxPut(pt.Bytes("box"), pt.Bytes("goonery")), pt.TealType.none),
+    (8, pt.BoxLen(pt.Bytes("box")), pt.TealType.none),
+    (8, pt.BoxGet(pt.Bytes("box")), pt.TealType.none),
+    (8, pt.BoxPut(pt.Bytes("box"), pt.Bytes("goonery")), pt.TealType.none),
+    (10, pt.BoxResize(pt.Bytes("box"), pt.Int(16)), pt.TealType.none),
+    (
+        10,
+        pt.BoxSplice(pt.Bytes("box"), pt.Int(5), pt.Int(2), pt.Bytes("replacement")),
+        pt.TealType.none,
+    ),
 ]
 
 
-@pytest.mark.parametrize("test_case, test_case_type", POSITIVE_TEST_CASES)
-def test_compile_version_and_type(test_case, test_case_type):
+@pytest.mark.parametrize("version, test_case, test_case_type", POSITIVE_TEST_CASES)
+def test_compile_version_and_type(version, test_case, test_case_type):
     with pytest.raises(pt.TealInputError):
-        test_case.__teal__(avm7Options)
+        test_case.__teal__(pt.CompileOptions(version=version - 1))
 
-    test_case.__teal__(avm8Options)
+    test_case.__teal__(pt.CompileOptions(version=version))
     assert test_case.type_of() == test_case_type
     assert not test_case.has_return()
-
-    return
 
 
 INVALID_TEST_CASES: list[Tuple[list[pt.Expr], type | Callable[..., pt.MaybeValue]]] = [
@@ -41,6 +47,12 @@ INVALID_TEST_CASES: list[Tuple[list[pt.Expr], type | Callable[..., pt.MaybeValue
     ([pt.Int(12)], pt.BoxLen),
     ([pt.Int(45)], pt.BoxGet),
     ([pt.Bytes("box"), pt.Int(123)], pt.BoxPut),
+    ([pt.Int(1), pt.Int(2)], pt.BoxResize),
+    ([pt.Bytes("box"), pt.Bytes("b")], pt.BoxResize),
+    ([pt.Bytes("box"), pt.Int(123), pt.Int(456), pt.Int(7)], pt.BoxSplice),
+    ([pt.Bytes("box"), pt.Int(123), pt.Bytes("456"), pt.Bytes("x")], pt.BoxSplice),
+    ([pt.Bytes("box"), pt.Bytes("123"), pt.Int(456), pt.Bytes("x")], pt.BoxSplice),
+    ([pt.Int(8), pt.Int(123), pt.Int(456), pt.Bytes("x")], pt.BoxSplice),
 ]
 
 
@@ -63,6 +75,25 @@ def test_box_create_compile():
         ]
     )
     actual, _ = expr.__teal__(avm8Options)
+    actual.addIncoming()
+    actual = pt.TealBlock.NormalizeBlocks(actual)
+
+    assert expected == actual
+
+
+def test_box_resize_compile():
+    name_arg: pt.Expr = pt.Bytes("eineName")
+    size_arg: pt.Expr = pt.Int(10)
+    expr: pt.Expr = pt.BoxResize(name_arg, size_arg)
+
+    expected = pt.TealSimpleBlock(
+        [
+            pt.TealOp(name_arg, pt.Op.byte, '"eineName"'),
+            pt.TealOp(size_arg, pt.Op.int, 10),
+            pt.TealOp(expr, pt.Op.box_resize),
+        ]
+    )
+    actual, _ = expr.__teal__(avm10Options)
     actual.addIncoming()
     actual = pt.TealBlock.NormalizeBlocks(actual)
 
@@ -119,6 +150,29 @@ def test_box_replace():
         ]
     )
     actual, _ = expr.__teal__(avm8Options)
+    actual.addIncoming()
+    actual = pt.TealBlock.NormalizeBlocks(actual)
+
+    assert expected == actual
+
+
+def test_box_splice():
+    name_arg: pt.Expr = pt.Bytes("eineName")
+    srt_arg: pt.Expr = pt.Int(10)
+    len_arg: pt.Expr = pt.Int(17)
+    replace_arg: pt.Expr = pt.Bytes("replace-str")
+    expr: pt.Expr = pt.BoxSplice(name_arg, srt_arg, len_arg, replace_arg)
+
+    expected = pt.TealSimpleBlock(
+        [
+            pt.TealOp(name_arg, pt.Op.byte, '"eineName"'),
+            pt.TealOp(srt_arg, pt.Op.int, 10),
+            pt.TealOp(len_arg, pt.Op.int, 17),
+            pt.TealOp(replace_arg, pt.Op.byte, '"replace-str"'),
+            pt.TealOp(expr, pt.Op.box_splice),
+        ]
+    )
+    actual, _ = expr.__teal__(avm10Options)
     actual.addIncoming()
     actual = pt.TealBlock.NormalizeBlocks(actual)